pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "PyMPDATA-MPI"
description = "PyMPDATA + numba-mpi coupler sandbox"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["MPI", "MPDATA", "Numba", "PyMPDATA"]
license = {text = "GPL-3.0"}
classifiers = [
        "Intended Audience :: Science/Research",
        "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3",
        "Topic :: Scientific/Engineering",
        "Topic :: Scientific/Engineering :: Atmospheric Science",
        "Topic :: Scientific/Engineering :: Mathematics",
        "Topic :: Scientific/Engineering :: Physics"
]
dependencies = [
    # TODO #71: these should be handled within PyMPDATA?
    "numba<0.57.0",
    "numpy<1.25.0",
    "numba_mpi>=0.30",
    "PyMPDATA==1.0.13",
    "mpi4py",
    "h5py",
]
dynamic = ["version"]

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-mpi",
    "pytest-timeout",
    "matplotlib",
]
examples = [
    "py-modelrunner",
]


